LPD-58579 Rename class and packages
