+++ tools/test-report-generator/src/main/java/com/liferay/test/report/generator/csv/playwright/PlaywrightTestSelector.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.test.report.generator.record.csv.playwright;
+package com.liferay.test.report.generator.csv.playwright;
 
 import java.io.IOException;
